Security problem fixed: database with password encryption
To prevent users of modifying the database externally. Also, new version
of database is used (Access 2003 to 2010 and .mbd to .accdb)
Also, I implemented a new functionality that allows the user to reset a
database or reset the whole program (including the settings)
Also, I fixed a bug in saving the bakcup while restoring an old backup.
Javadoc updated
Jar updated